Plan A Trip
The project's purpose is to obtain information about people who have come from abroad and Turkey's ability to plan trips in a relaxed way in Turkey. The homepage of my project includes the introduction post, the last 3 blogs and the description and the read more button, the places I like and the last 10 blogs. Cities are part of Turkey's cities. Inside each city there are descriptions, restaurants, drinks, things to do, and car rental buttons. In the blogs section, there are blogs written by users and there are explanations and comments can be made on the blogs. In the About section, what the site does is explained. You can contact me from the contact section. Finally, you can register and log in to the site in the login section. Registered users on the site can post blogs. We can view, update, and delete blogs written from the admin panel. We can see, update and delete comments written. We can see and delete the last communication posts from the admin panel. :balloon: